requirements.txt
aiohttp~=3.8.3
aiologger~=0.7.0
python-dotenv~=0.21.0
